# backend/docs/swagger.yaml
definitions:
...
  domain.FetchLevelStruct:
    properties:
      filter_params:
        $ref: '#/definitions/domain.LevelFilterParams'
      page_info:
        $ref: '#/definitions/github_com_Gadzet005_GoType_backend_internal_domain_Level.PageInfo'
      sort_params:
        $ref: '#/definitions/domain.LevelSortParams'
      tags:
        items:
          type: string
        type: array
    required:
    - filter_params
    - page_info
    - sort_params
    - tags
    type: object
  domain.GetLevelInfoStruct:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  domain.Level:
    properties:
      author:
        type: integer
      author_name:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      duration:
        type: integer
      id:
        type: integer
      image_type:
        type: string
      language:
        type: string
      name:
        type: string
      preview_path:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
    required:
    - author
    - author_name
    - description
    - difficulty
    - duration
    - image_type
    - language
    - name
    - tags
    - type
    type: object
...
  domain.LevelFilterParams:
    properties:
      difficulty:
        type: integer
      language:
        type: string
      level_name:
        type: string
    type: object
  domain.LevelInfo:
    properties:
      level_info:
        $ref: '#/definitions/domain.Level'
    type: object
  domain.LevelSortParams:
    properties:
      date:
        type: string
      popularity:
        type: string
    type: object
  domain.LevelsList:
    properties:
      levels:
        items:
          $ref: '#/definitions/domain.Level'
        type: array
    type: object
...
  domain.RefreshStruct:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  domain.User:
    properties:
      name:
        type: string
      password:
        description: hash
        type: string
    required:
    - name
    - password
    type: object
...
  github_com_Gadzet005_GoType_backend_internal_domain_Level.PageInfo:
    properties:
      offset:
        type: integer
      page_size:
        type: integer
    required:
    - offset
    - page_size
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for GoType game and website
  title: GoType App API
  version: 0.0.1
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: authorize using login and password
      operationId: login
      parameters:
      - description: login and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.RefreshStruct'
        "400":
          description: 'Possible messages: ERR_INVALID_INPUT - Wrong structure of
            input json; ERR_NO_SUCH_USER - User with such name and password does not
            exist;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server; '
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: get new access token and refresh token by existing access token
        and refresh token
      operationId: refresh
      parameters:
      - description: RefreshToken and AccessToken
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.RefreshStruct'
        "400":
          description: 'Possible messages: ERR_INVALID_INPUT - Wrong structure of
            input json; ERR_NO_SUCH_USER - User with id as in access token does not
            exist; ERR_ACCESS_TOKEN_WRONG - Wrong access token; ERR_REFRESH_TOKEN_WRONG
            - Wrong refresh token;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Refresh token expired;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server; '
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Refresh
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: create new account
      operationId: create-account
      parameters:
      - description: new account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.RefreshStruct'
        "400":
          description: 'Possible messages: ERR_INVALID_INPUT - Wrong structure of
            input json; ERR_USER_EXISTS - User with such name already exists;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server; '
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Register
      tags:
      - auth
  /level/create-level:
    post:
      consumes:
      - multipart/form-data
      description: Create level with given structure
      operationId: create-level
      parameters:
      - description: Archive with level.
        in: formData
        name: level
        required: true
        type: file
      - description: JSON file with level description.
        in: formData
        name: info
        required: true
        type: file
      - description: File with preview image of the level
        in: formData
        name: preview
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetLevelInfoStruct'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create level
      tags:
      - level
  /level/download-level:
    get:
      consumes:
      - application/json
      description: Download level with given id from server
      operationId: download-level
      parameters:
      - description: id of level you want to download
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.GetLevelInfoStruct'
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Archive with level.
          schema:
            type: file
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json; ERR_ENTITY_NOT_FOUND - no such level on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Download level
      tags:
      - level
  /level/get-level-info:
    get:
      consumes:
      - application/json
      description: Get level info about level with given id
      operationId: get-level-info
      parameters:
      - description: id of level you want to find out about
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.GetLevelInfoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LevelInfo'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get level info
      tags:
      - level
  /level/get-level-list:
    get:
      consumes:
      - application/json
      description: Get level list with given params
      operationId: get-level-list
      parameters:
      - description: search and filter params
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.FetchLevelStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LevelsList'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get level list
      tags:
      - level
  /level/update-level:
    post:
      consumes:
      - multipart/form-data
      description: Update level with given structure
      operationId: update-level
      parameters:
      - description: Archive with level.
        in: formData
        name: level
        required: true
        type: file
      - description: JSON file with level description.
        in: formData
        name: info
        required: true
        type: file
      - description: File with preview image of the level
        in: formData
        name: preview
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetLevelInfoStruct'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_INVALID_INPUT - Wrong structure of
            input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update level
      tags:
      - level
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"